% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.R
\name{teachrag_deps}
\alias{teachrag_deps}
\title{Required packages for teachrag}
\usage{
teachrag_deps()
}
\value{
A data frame with package names and minimum versions.
}
\description{
Required packages for teachrag
}
\keyword{internal}
